single product component rendering on all pages view
