+++ client/components/products/all-products.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
+import SingleProduct from './single-product' 
 
 //--------------Component-------------------
 class Products extends Component {
@@ -49,23 +50,7 @@ class Products extends Component {
         </h2>
         <div className="parent-product-container">
           {filteredProducts.length ? (
-            filteredProducts.map(product => {
-              let productImage = {
-                backgroundImage: `url('${product.imgUrl}')`
-              }
-              return (
-                <div key={product.id} className="product-container">
-                  <div className="product-name">{product.name}</div>
-                  <div className="product-imgUrl" style={productImage} />
-                  <div className="product-description">
-                    {product.description}
-                  </div>
-                  <div className="product-rating">{product.rating}</div>
-                  <div className="product-price">{product.price}</div>
-                  <button onSubmit={() => updateCart(product)} type="submit" className="product-add-to-cart">Add to Cart</button>
-                </div> 
-              )
-            })
+            filteredProducts.map(product => <SingleProduct product={product} key={product.id} />)
           ) : (
             <h2>No products available with that rating.</h2>
           )}
